fix(discover): stop throwing 503 item render
This was a legacy of server side renders ... and even then it was a horrid error.  This is a bit more graceful until we have a potential empty state.  If something is truly wrong here, we will get an error, but at least this lets the user navigate away.  Ideally we have a specific empty state for this.